change button implementation (code is broken)
